seperate UserSettingStore from CurrentUser
